Relocate example script utilities to improve structure
- Move example script utilities:
  - `paper_reproduction_utils.py` from `examples/lti_control/` to
    `examples/lti_control/reproduction_utilities/`.
  - `nonlinear_cstr_model.py` from `examples/nonlinear_control/`
    to `examples/nonlinear_control/utilities/`.
- Update `README.md` to align with these changes.

This change improves code structure by separating utility modules
from example scripts.
