+++ examples/nonlinear_control/utilities/nonlinear_cstr_model.py
